cloud/aws: Use the default credential constants as the last option for aws credential provider
