+++ pkg/cloud/aws/credentials_default.go
@@ -4,6 +4,7 @@
 package aws
 
 import (
+	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go/aws/credentials"
 )
 
@@ -18,5 +19,9 @@ const (
 // tests, you get this implementation that tells the AWS SDK to use the default credentials (as if you didn't specify
 // any credentials at all).
 func GetDefaultCredentials() *credentials.Credentials {
+	return nil
+}
+
+func GetDefaultProvider() aws.CredentialsProvider {
 	return nil
 }
